作業途中コミット　動作確認済み
